% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate Euclidean distance between two vectors.}
\usage{
calculate_distance(x, y)
}
\arguments{
\item{x, y}{Numeric vectors.}
}
\value{
The distance between the vectors.
}
\description{
Calculate Euclidean distance between two vectors.
}
\examples{
x <- c(2, 5)
y <- c(5, 7)
calculate_distance(x, y)
}
\author{
Antoni Baum \email{antoni.baum@protonmail.com}
}
