% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_methods.R
\name{group_by_average}
\alias{group_by_average}
\title{Group linearly ordered alternatives by average.}
\usage{
group_by_average(ordered_df)
}
\arguments{
\item{ordered_df}{A data frame with elements \code{alt.row}, \code{score}
and \code{rank} (returned from any linear ordering method from this library).}
}
\value{
A list including elements:
\describe{
  \item{\code{highest}}
  \item{\code{above_average}}
  \item{\code{below_average}}
  \item{\code{lowest}}
}
}
\description{
Group linearly ordered alternatives by average.
}
\examples{
d <- matrix(rpois(12, 5), nrow = 3, ncol = 3)
w <- c(1, 1, 2)
i <- c('+', '-', '+')
group_by_average(sum_of_ranks(d, w, i))
}
\author{
Antoni Baum \email{antoni.baum@protonmail.com}
}
