% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping_methods.R
\name{group_by_quartile}
\alias{group_by_quartile}
\title{Group linearly ordered alternatives by quartiles.}
\usage{
group_by_quartile(ordered_df)
}
\arguments{
\item{ordered_df}{A data frame with elements \code{alt.row}, \code{score}
and \code{rank} (returned from any linear ordering method from this library).}
}
\value{
A list including elements:
\describe{
  \item{\code{first_quartile}}{group of values in first quartile}
  \item{\code{second_quartile}}{group of values in second quartile}
  \item{\code{third_quartile}}{group of values in third quartile}
  \item{\code{fourth_quartile}}{group of values in fourth quartile}
}
}
\description{
Group linearly ordered alternatives by quartiles.
}
\examples{
d <- matrix(rpois(12, 5), nrow = 3, ncol = 3)
w <- c(1, 1, 2)
i <- c('+', '-', '+')
group_by_quartile(sum_of_ranks(d, w, i))
}
\author{
Antoni Baum \email{antoni.baum@protonmail.com}
}
