% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{normalize_impacts}
\alias{normalize_impacts}
\title{Normalize impacts to all positive by replacing negative influences
with their negatives.}
\usage{
normalize_impacts(decision, impacts)
}
\arguments{
\item{decision}{A numeric matrix with m rows for m alternatives and n columns
for n criterions.}

\item{impacts}{A character vector of '+' and '-' signs for the way that each
criterion influences on the alternatives.}
}
\value{
A copy of \code{decision} matrix with normalized impacts.
}
\description{
Normalize impacts to all positive by replacing negative influences
with their negatives.
}
\examples{
d <- matrix(rpois(12, 5), nrow = 3, ncol = 3)
i <- c('+', '-', '+')
normalize_impacts(d, i)
}
\author{
Antoni Baum \email{antoni.baum@protonmail.com}
}
