% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{topsis_normalize}
\alias{topsis_normalize}
\title{Normalization for TOPSIS. Intended to be used with \code{apply} function
on a target matrix.}
\usage{
topsis_normalize(decision, weights)
}
\arguments{
\item{decision}{A numeric matrix with m rows for m alternatives and n columns
for n criterions.}

\item{weights}{A numeric vector with length equal to number of columns in
decision matrix for weights of criterions.}
}
\value{
A copy of \code{decision} matrix with TOPSIS-normalized values.
}
\description{
Normalization for TOPSIS. Intended to be used with \code{apply} function
on a target matrix.
}
\examples{
d <- matrix(rpois(12, 5), nrow = 3, ncol = 3)
w <- c(1, 1, 2)
d <- sapply(d, topsis_normalize, w)
}
\references{
Hwang, C.L.; Yoon, K. (1981). Multiple Attribute Decision
Making: Methods and Applications. New York: Springer-Verlag.
}
\author{
Antoni Baum \email{antoni.baum@protonmail.com}
}
